backend/requirements.txt
uvicorn==0.34.0
fastapi==0.115.11
sqladmin==0.20.1
python-jose==3.4.0
bcrypt==4.3.0
pydantic-settings==2.8.1
fastapi-storages==0.3.0
pytz==2025.1
email_validator==2.2.0
aiosqlite==0.21.0
elasticsearch==8.17.2
aiohttp==3.11.14
uuid6==2024.7.10
alembic==1.15.1
asyncpg==0.30.0
psycopg==3.2.6
python-json-logger==3.3.0
prometheus-fastapi-instrumentator==7.1.0
celery==5.5.0
redis==5.2.1
asgiref==3.8.1
pytest==8.3.5
pytest-asyncio==0.26.0
httpx==0.28.1
cryptography==44.0.2
PyJWT